more changes in JS using DATATABLE e JQUERY

+++ assets/js/aula2.js
@@ -1,4 +1,14 @@
-// $("#botao").addEventListener("click", function () {
+$(document).ready(function(){
+    consomeAPI();
+    $("#tabela").DataTable();
+});
+
+$("#botao").click(function () {
+    alert("teste");
+});
+
+
+// document.getElementById("botao").addEventListener("click", function () {
 //     alert("teste");
 // });
 
@@ -13,14 +23,22 @@ function consomeAPI(){
 		url:'http://formtalentosapi.azurewebsites.net/api/Episodio',
 		type:'GET',
 		dataType:'json',
-		crossDomain: true,
+        crossDomain: true,
+        async: false,
         success: retSucesso,
         error: retErro
-    });
+    })
 }
 
 function limpaTabela(){
     $("#tabela tbody").empty();
+
+    $("#id").val('');
+    $("#dataLancamento").val('');
+    $("#descricao").val('');
+    $("#nomeEpisodio").val('');
+    $("#numeroDoEp").val('');
+    $("#temporada").val('');  
 }
 
 function pegaDados(){
@@ -44,22 +62,27 @@ function inserirAPI(){
 		dataType:'json',
         crossDomain: true,
         data: ret,
-        success: alert('sucesso'),
+        success: alert('registro inserido com sucesso!'),
         error: retErro
+    }).done(function(){
+        consomeAPI();
     });
+
 }
 
-function deletaAPI(){
-    var id = $("#id").val()
-
-    $.ajax({
-		url:'http://formtalentosapi.azurewebsites.net/api/Episodio/' + id,
-		type:'DELETE',
-		dataType:'json',
-        crossDomain: true,
-        success: alert('sucesso'),
-        error: retErro
-    });
+function deletaAPI(id){
+    if (confirm('você tem certeza que deseja deletar?')){
+        $.ajax({
+            url:'http://formtalentosapi.azurewebsites.net/api/Episodio/' + id,
+            type:'DELETE',
+            dataType:'json',
+            crossDomain: true,
+            success: alert('registro deletado com sucesso!'),
+            error: retErro
+        }).done(function(data){
+            consomeAPI();
+        });
+    }
 }
 
 function atualizaAPI(){
@@ -71,15 +94,19 @@ function atualizaAPI(){
 		dataType:'json',
         crossDomain: true,
         data: ret,
-        success: alert('sucesso'),
+        success: alert('registro atualizado com sucesso!'),
         error: retErro
+    }).done(function(){
+        consomeAPI();
     });
+
 }
 
 function retSucesso(data){
     data.forEach(function(element) {
         $("#tabela tbody").append(`<tr><td>${element.id}</td><td>${element.dataLancamento}</td><td>${element.descricao}</td>
-        <td>${element.nomeEpisodio}</td><td>${element.numeroDoEp}</td><td>${element.temporada}</td><tr>`);
+        <td>${element.nomeEpisodio}</td><td>${element.numeroDoEp}</td><td>${element.temporada}</td><td>
+        <img src="../assets/img/delete.png" onclick="deletaAPI(${element.id})" style="cursor: pointer; margin: none !important" /></td></tr>`);
     });
 }
